remove AgendaItem in R
