+++ packages/server/dataloader/rethinkPrimaryKeyLoaderMakers.ts
@@ -3,7 +3,6 @@ import RethinkPrimaryKeyLoaderMaker from './RethinkPrimaryKeyLoaderMaker'
 /**
  * all rethink dataloader types which also must exist in {@link rethinkDriver/RethinkSchema}
  */
-export const agendaItems = new RethinkPrimaryKeyLoaderMaker('AgendaItem')
 export const comments = new RethinkPrimaryKeyLoaderMaker('Comment')
 export const reflectPrompts = new RethinkPrimaryKeyLoaderMaker('ReflectPrompt')
 export const massInvitations = new RethinkPrimaryKeyLoaderMaker('MassInvitation')
