+++ packages/server/graphql/mutations/helpers/addAgendaItemToActiveActionMeeting.ts
@@ -47,7 +47,6 @@ const addAgendaItemToActiveActionMeeting = async (
         updatedAt: now
       })
       .run(),
-    r.table('AgendaItem').get(agendaItemId).update({meetingId: meetingId}).run(),
     getKysely().updateTable('AgendaItem').set({meetingId}).where('id', '=', agendaItemId).execute(),
     insertDiscussions([
       {
