--- packages/server/graphql/mutations/updateAgendaItem.ts
@@ -1,111 +1,0 @@
-import {GraphQLNonNull} from 'graphql'
-import {SubscriptionChannel} from 'parabol-client/types/constEnums'
-import makeUpdateAgendaItemSchema from 'parabol-client/validation/makeUpdateAgendaItemSchema'
-import {getSortOrder} from '../../../client/shared/sortOrder'
-import getRethink from '../../database/rethinkDriver'
-import AgendaItemsStage from '../../database/types/AgendaItemsStage'
-import getKysely from '../../postgres/getKysely'
-import {getUserId, isTeamMember} from '../../utils/authorization'
-import getPhase from '../../utils/getPhase'
-import publish from '../../utils/publish'
-import standardError from '../../utils/standardError'
-import {GQLContext} from '../graphql'
-import UpdateAgendaItemInput, {UpdateAgendaItemInputType} from '../types/UpdateAgendaItemInput'
-import UpdateAgendaItemPayload from '../types/UpdateAgendaItemPayload'
-
-export default {
-  type: UpdateAgendaItemPayload,
-  description: 'Update an agenda item',
-  args: {
-    updatedAgendaItem: {
-      type: new GraphQLNonNull(UpdateAgendaItemInput),
-      description: 'The updated item including an id, content, status, sortOrder'
-    }
-  },
-  async resolve(
-    _source: unknown,
-    {updatedAgendaItem}: {updatedAgendaItem: UpdateAgendaItemInputType},
-    {authToken, dataLoader, socketId: mutatorId}: GQLContext
-  ) {
-    const now = new Date()
-    const r = await getRethink()
-    const pg = getKysely()
-    const operationId = dataLoader.share()
-    const subOptions = {mutatorId, operationId}
-    const viewerId = getUserId(authToken)
-
-    // AUTH
-    const {id: agendaItemId} = updatedAgendaItem
-    const [teamId] = agendaItemId.split('::') as [string]
-    if (!isTeamMember(authToken, teamId)) {
-      return standardError(new Error('Team not found'), {userId: viewerId})
-    }
-
-    // VALIDATION
-    const schema = makeUpdateAgendaItemSchema()
-    const {
-      errors,
-      data: {id, ...doc}
-    } = schema(updatedAgendaItem) as any
-    if (Object.keys(errors).length) {
-      return standardError(new Error('Failed input validation'), {userId: viewerId})
-    }
-
-    // RESOLUTION
-    const oldAgendaItems = await dataLoader.get('agendaItemsByTeamId').load(teamId)
-    const fromIdx = oldAgendaItems.findIndex((agendaItem) => agendaItem.id === id)
-    await r
-      .table('AgendaItem')
-      .get(id)
-      .update({
-        ...doc,
-        updatedAt: now
-      })
-      .run()
-    dataLoader.clearAll('agendaItems')
-    if (doc.sortOrder !== null && doc.sortOrder !== undefined) {
-      const nextAgendaItems = await dataLoader.get('agendaItemsByTeamId').load(teamId)
-      const pgagendaItems = await dataLoader.get('_pgagendaItemsByTeamId').load(teamId)
-      const toIdx = nextAgendaItems.findIndex((agendaItem) => agendaItem.id === id)
-      const pgSortOrder = getSortOrder(pgagendaItems, fromIdx, toIdx)
-      await pg
-        .updateTable('AgendaItem')
-        .set({sortOrder: pgSortOrder})
-        .where('id', '=', id)
-        .execute()
-    } else {
-      await pg
-        .updateTable('AgendaItem')
-        .set({pinned: doc.pinned, content: doc.content})
-        .where('id', '=', id)
-        .execute()
-    }
-
-    const activeMeetings = await dataLoader.get('activeMeetingsByTeamId').load(teamId)
-    const actionMeeting = activeMeetings.find(
-      (activeMeeting) => activeMeeting.meetingType === 'action'
-    )
-    const meetingId = actionMeeting?.id ?? null
-    if (actionMeeting) {
-      const {id: meetingId, phases} = actionMeeting
-      const agendaItemPhase = getPhase(phases, 'agendaitems')
-      const {stages} = agendaItemPhase
-      const agendaItems = await dataLoader.get('agendaItemsByTeamId').load(teamId)
-      const getSortOrder = (stage: AgendaItemsStage) => {
-        const agendaItem = agendaItems.find((item) => item.id === stage.agendaItemId)
-        return (agendaItem && agendaItem.sortOrder) || 0
-      }
-      stages.sort((a, b) => (getSortOrder(a) > getSortOrder(b) ? 1 : -1))
-      await r
-        .table('NewMeeting')
-        .get(meetingId)
-        .update({
-          phases
-        })
-        .run()
-    }
-    const data = {agendaItemId, meetingId}
-    publish(SubscriptionChannel.TEAM, teamId, 'UpdateAgendaItemPayload', data, subOptions)
-    return data
-  }
-}
