+++ packages/server/graphql/public/typeDefs/AgendaItem.graphql
@@ -40,7 +40,7 @@ type AgendaItem {
   """
   The sort order of the agenda item in the list
   """
-  sortOrder: Float!
+  sortOrder: String!
 
   """
   *The team for this agenda item
