+++ packages/server/graphql/public/typeDefs/CreateAgendaItemInput.graphql
@@ -18,7 +18,7 @@ input CreateAgendaItemInput {
   """
   The sort order of the agenda item in the list
   """
-  sortOrder: Float
+  sortOrder: String
 
   """
   The meeting ID of the agenda item
