+++ packages/server/graphql/public/types/AddAgendaItemPayload.ts
@@ -1,0 +1,24 @@
+import {AddAgendaItemPayloadResolvers} from '../resolverTypes'
+
+export type AddAgendaItemPayloadSource =
+  | {
+      agendaItemId: string
+      meetingId?: string
+    }
+  | {error: {message: string}}
+
+const AddAgendaItemPayload: AddAgendaItemPayloadResolvers = {
+  agendaItem: (source, _args, {dataLoader}) => {
+    return 'agendaItemId' in source
+      ? dataLoader.get('agendaItems').loadNonNull(source.agendaItemId)
+      : null
+  },
+
+  meeting: (source, _args, {dataLoader}) => {
+    return 'meetingId' in source && source.meetingId
+      ? dataLoader.get('newMeetings').load(source.meetingId)
+      : null
+  }
+}
+
+export default AddAgendaItemPayload
