+++ packages/server/graphql/public/types/AgendaItemsStage.ts
@@ -3,7 +3,7 @@ import {AgendaItemsStageResolvers} from '../resolverTypes'
 const AgendaItemsStage: AgendaItemsStageResolvers = {
   __isTypeOf: ({phaseType}) => phaseType === 'agendaitems',
   agendaItem: ({agendaItemId}, _args, {dataLoader}) => {
-    return dataLoader.get('agendaItems').load(agendaItemId)
+    return dataLoader.get('agendaItems').loadNonNull(agendaItemId)
   }
 }
